Version 1.1 Replace height and weight sliders with textfields.

+++ BMI Calculator Practice/Model/BMI.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 struct BMI {
-    let value: Float
+    let value: Double
     let advice: String
     let color: UIColor
 }
